Auto tasks & cspell configuration
If in the project directory or the current directory theres a file under
.nvim/auto-tasks.lua then for each entry a terminal is spawned. Also a
keymap can be assigned explicitly. if no keymap is provided the
<A-[1-9]> terminals will be used to spawn the tasks.

The cspell config file is stored under project dir in .nvim folder if
that exists in the current directory in .nvim folder if that exists in
nvim config directory if none of above exists.